+++ src/Summoner/Decision.hs
@@ -8,7 +8,7 @@ module Summoner.Decision
 import Generics.Deriving.Monoid (GMonoid (..))
 import Generics.Deriving.Semigroup (GSemigroup (..))
 
-import Summoner.Question (chooseYesNoBool, falseMessage, trueMessage)
+import Summoner.Question (YesNoPrompt (..), chooseYesNoBool, falseMessage, trueMessage)
 
 
 -- | Used for detecting the user decision during CLI input.
@@ -32,8 +32,8 @@ instance GMonoid Decision where
     gmempty = mempty
     gmappend = (<>)
 
-decisionToBool :: Decision -> Text -> IO Bool
-decisionToBool decision target = case decision of
-    Yes -> trueMessage  target
-    Nop -> falseMessage target
+decisionToBool :: Decision -> YesNoPrompt -> IO Bool
+decisionToBool decision target@YesNoPrompt {..} = case decision of
+    Yes -> trueMessage  yesNoTarget
+    Nop -> falseMessage yesNoTarget
     Idk -> chooseYesNoBool target